+++ src/layers/GYScreenshotPopup.hpp
@@ -2,6 +2,7 @@
 
 #include <Geode/Geode.hpp>
 #include <Geode/modify/FLAlertLayer.hpp>
+#include <Geode/utils/web.hpp>
 #include <string>
 
 #include "../Tags.hpp"
@@ -10,6 +11,10 @@ using namespace geode::prelude;
 
 class GYScreenshotPopup : public geode::Popup<int const&> {
 protected:
+    EventListener<web::WebTask> m_downloadListener;
+    Ref<CCImage> m_image;
+    CCSprite* m_sprite;
+
     bool setup(int const& layer) override {
         Tags tags;
         auto layerName = tags.getStringFromTag(layer);
@@ -31,32 +36,58 @@ protected:
         m_mainLayer->setContentSize({ winSize.width * 0.6f, winSize.height * 0.7f });
         m_mainLayer->updateLayout();
 
-        /*auto sprite = CCSprite::createWithSpriteFrameName(fmt::format("{}Preview.png"_spr, layerName).c_str());
-        if (!sprite) {
-            log::error("Sprite not found for frame: {}", fmt::format("{}Preview.png", layerName));
-            return false;
-        }
-
-        log::debug("Loading sprite: {}", fmt::format("omgrod.geodify/{}Preview.png", layerName));
-
-        auto maxWidth = m_mainLayer->getContentSize().width * 0.75f;
-        auto maxHeight = m_mainLayer->getContentSize().height * 0.75f;
-
-        auto spriteWidth = sprite->getContentSize().width;
-        auto spriteHeight = sprite->getContentSize().height;
-
-        float scale = std::min(maxWidth / spriteWidth, maxHeight / spriteHeight);
-        sprite->setScale(scale);
-
-        m_mainLayer->addChildAtPosition(sprite, Anchor::Center);*/
-
-        auto warning = CCLabelBMFont::create("This feature is not yet implemented. Check back later.", "bigFont.fnt");
-        warning->setPosition({ winSize.width * .3f, winSize.height * .35f });
-        warning->setScale(0.3f);
-
-        m_mainLayer->addChild(warning);
+        // Construct URL for the raw GitHub image
+        std::string URL = fmt::format("https://raw.githubusercontent.com/OmgRod/Geodify/master/previews/{}Preview.png", layerName);
+        
+        auto req = web::WebRequest();
+        m_downloadListener.bind([this](web::WebTask::Event* e) {
+            if (auto res = e->getValue()) {
+                if (!res->ok()) {
+                    log::error("Failed to download image from URL.");
+                    onDownloadFail();
+                } else {
+                    auto data = res->data();
+                    std::thread imageThread = std::thread([data, this]() {
+                        m_image = new CCImage();
+                        m_image->autorelease();
+                        m_image->initWithImageData(const_cast<uint8_t*>(data.data()), data.size());
+                        geode::Loader::get()->queueInMainThread([this]() {
+                            imageCreationFinished(m_image);
+                        });
+                    });
+                    imageThread.detach();
+                }
+            }
+        });
+        m_downloadListener.setFilter(req.get(URL));
 
         return true;
+    }
+
+    void imageCreationFinished(CCImage* image) {
+        CCTexture2D* texture = new CCTexture2D();
+        texture->autorelease();
+        texture->initWithImage(image);
+
+        m_sprite = CCSprite::createWithTexture(texture);
+        if (m_sprite) {
+            auto maxWidth = m_mainLayer->getContentSize().width * 0.75f;
+            auto maxHeight = m_mainLayer->getContentSize().height * 0.75f;
+
+            auto spriteWidth = m_sprite->getContentSize().width;
+            auto spriteHeight = m_sprite->getContentSize().height;
+
+            float scale = std::min(maxWidth / spriteWidth, maxHeight / spriteHeight);
+            m_sprite->setScale(scale);
+
+            m_mainLayer->addChildAtPosition(m_sprite, Anchor::Center);
+        }
+    }
+
+    void onDownloadFail() {
+        // Handle failure if the image could not be downloaded
+        CCSprite* fallbackSprite = CCSprite::create("noPreview.png"_spr);
+        m_mainLayer->addChildAtPosition(fallbackSprite, Anchor::Center);
     }
 
 public:
@@ -67,6 +98,7 @@ public:
         }
         return input;
     }
+
     static GYScreenshotPopup* create(int const& text) {
         auto ret = new GYScreenshotPopup();
         auto winSize = CCDirector::sharedDirector()->getWinSize();